Worked on errors from txt file

+++ src/events/events.service.ts
@@ -1,24 +1,24 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { Prisma } from '@prisma/client';
+import { EventFilterDto } from './dto';
 
 @Injectable()
 export class EventsService {
   constructor(private prisma: PrismaService) {}
 
-  async create(createEventDto: Prisma.EventCreateInput, hostId: number) {
+  async create(createEventDto: Prisma.EventCreateInput, hostId: string) {
+    const { role, email } = await this.prisma.user.findUnique({
+      where: {
+        id: hostId
+      }
+    })
+
+    if(!email) throw new HttpException("User does not exist", HttpStatus.FORBIDDEN)
+    if(role !== "HOST") throw new HttpException("User is not allowed to create an event", HttpStatus.FORBIDDEN)
+      
     try {
-      const {
-        eventName,
-        date,
-        location,
-        type,
-        time,
-        prices,
-        bgImage,
-        noTickets,
-        ticketImage,
-      } = createEventDto;
+      const { eventName } = createEventDto;
       const event = await this.prisma.event.findUnique({
         where: {
           eventName,
@@ -32,19 +32,12 @@ export class EventsService {
         );
       const newEvent = await this.prisma.event.create({
         data: {
-          bgImage,
-          date,
-          location,
-          ticketImage,
-          eventName,
-          time,
-          prices,
-          type,
-          noTickets,
-          hostId,
-        },
-        include: {
-          host: true,
+          ...createEventDto,
+          host: {
+            connect: {
+              id: hostId
+            }
+          },
         },
       });
       return newEvent;
@@ -60,18 +53,114 @@ export class EventsService {
     }
   }
 
-  async findAll() {
+  async findAll(query: EventFilterDto) {
+    const { date, name, type, page, size, status } = query
+    const take = (size && page) ? Number(size) : undefined
+    const skip = (size && page) ? Number(page - 1) * take : undefined
+
+    const whereClause = {
+      AND: [],
+    }
+
+    if(name){
+      whereClause.AND.push({
+        eventName: {
+          contains: name,
+          mode: "insensitive"
+        }
+      })
+    }
+
+    if(type){
+      whereClause.AND.push({
+        type
+      })
+    }
+
+    if(date){
+      whereClause.AND.push({
+        date: {
+          has: date ?? null
+        }
+      })
+    }
+
+    if(status){
+      whereClause.AND.push({
+        status
+      })
+    }
+
+    
+
     const events = await this.prisma.event.findMany({
-      relationLoadStrategy: 'join', // or 'query'
+      where: whereClause,
+      take,
+      skip,
       include: {
         tickets: true,
-        host: true,
-      },
+        host: true
+      }
     });
     return events;
   }
 
-  async findOne(id: number) {
+  async findAllByCreator(id: string, query: EventFilterDto) {
+
+    if(id.length !== 24) throw new HttpException("User does not exist", HttpStatus.UNAUTHORIZED)
+    const user = await this.prisma.user.findUnique({
+      where: { id }
+    })
+    if(!user) throw new HttpException("User does not exist", HttpStatus.UNAUTHORIZED)
+    
+    const { date, name, type, page, size } = query
+    const take = (size && page) ? Number(size) : undefined
+    const skip = (size && page) ? Number(page - 1) * take : undefined
+    
+    const whereClause = {
+      AND: [],
+    }
+
+    if(name){
+      whereClause.AND.push({
+        eventName: {
+          contains: name,
+          mode: "insensitive"
+        }
+      })
+    }
+
+    if(type){
+      whereClause.AND.push({
+        type
+      })
+    }
+
+    if(date){
+      whereClause.AND.push({
+        date: {
+          has: date ?? null
+        }
+      })
+    }
+
+    const events = await this.prisma.event.findMany({
+      where: {
+        ...whereClause,
+        hostId: id
+      },
+      take,
+      skip,
+      include: {
+        tickets: true,
+        host: true
+      }
+    });
+    return events;
+  }
+
+  async findOne(id: string) {
+    if(id.length !== 24) throw new HttpException("Event does not exist", HttpStatus.NOT_FOUND)
     const event = await this.prisma.event.findUnique({
       where: {
         id,
@@ -83,18 +172,8 @@ export class EventsService {
     return event;
   }
 
-  async searchEvents(query: string) {
-    const events = await this.prisma.event.findMany({
-      where: {
-        eventName: {
-          contains: query,
-        },
-      },
-    });
-    return events;
-  }
-
-  async update(id: number, updateEventDto: Prisma.EventUpdateInput) {
+  async update(id: string, updateEventDto: Prisma.EventUpdateInput) {
+    if(id.length !== 24) throw new HttpException("Event does not exist", HttpStatus.NOT_FOUND)
     const event = await this.findOne(id);
 
     if (!event)
@@ -102,21 +181,29 @@ export class EventsService {
         'This event does not exist',
         HttpStatus.NOT_FOUND,
       );
-
-    await this.prisma.event.update({
-      where: { id },
-      data: updateEventDto,
-    });
-
-    return `This action updates a #${id} event`;
-  }
-
-  async remove(id: number) {
-    const event = await this.prisma.event.delete({
-      where: {
-        id,
-      },
-    });
-    return `This action removes a #${id} event ${event.eventName}`;
+    
+      await this.prisma.event.update({
+        where: { id },
+        data: updateEventDto,
+      });
+      
+      return { success: true, message: "Event Updated Successfully" }
+    }
+    
+    async remove(id: string) {
+      if(id.length !== 24) throw new HttpException("Event does not exist", HttpStatus.NOT_FOUND)
+      const event = await this.findOne(id);
+      if (!event)
+        throw new HttpException(
+          'This event does not exist',
+          HttpStatus.NOT_FOUND,
+        );
+      
+      await this.prisma.event.delete({
+        where: {
+          id,
+        },
+      });
+      return { success: true, message: "Event deleted Successfully" }
   }
 }
